route refactoring

+++ app/Http/Controllers/Tenant/CourseController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers\Tenant;
 
+use App\{Course, BranchOffice};
 use App\Http\Controllers\Controller;
-use App\{Course, Institute, TypeCourse};
+use Symfony\Component\HttpFoundation\Response;
 use App\Http\Requests\Tenant\StoreCourseRequest;
 use App\Http\Requests\Tenant\UpdateCourseRequest;
 
@@ -18,78 +19,81 @@ class CourseController extends Controller
     }
 
     /**
-     * @param \App\Institute $institute
+     * @param \App\BranchOffice $branchOffice
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function index(Institute $institute)
+    public function index(BranchOffice $branchOffice)
     {
         $this->authorize('tenant-view', Course::class);
-        $courses = Course::with('typecourse')->paginate();
-        return view('tenant.course.index', compact('institute', 'courses'));
+        $courses = $branchOffice->courses()->paginate();
+        return view('tenant.course.index', compact('branchOffice', 'courses'));
     }
 
     /**
-     * @param \App\Institute $institute
+     * @param \App\BranchOffice $branchOffice
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function create(Institute $institute)
+    public function create(BranchOffice $branchOffice)
     {
         $this->authorize('tenant-create', Course::class);
-        $typeCourses = TypeCourse::all();
-        return view('tenant.course.create', compact('institute', 'typeCourses'));
+        $typeCourses = $branchOffice->typecourses;
+        return view('tenant.course.create', compact('branchOffice', 'typeCourses'));
     }
 
     /**
      * @param \App\Http\Requests\Tenant\StoreCourseRequest $request
-     * @param \App\Institute $institute
+     * @param \App\BranchOffice $branchOffice
      * @return \Illuminate\Http\RedirectResponse
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function store(StoreCourseRequest $request, Institute $institute)
+    public function store(StoreCourseRequest $request, BranchOffice $branchOffice)
     {
         $this->authorize('tenant-create', Course::class);
-        return redirect()->route('tenant.courses.index', $institute)->with(['flash_success' => $request->createCourse()]);
+        return redirect()->route('tenant.courses.index', $branchOffice)->with(['flash_success' => $request->createCourse($branchOffice)]);
     }
 
     /**
-     * @param \App\Institute $institute
+     * @param \App\BranchOffice $branchOffice
      * @param \App\Course $course
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function edit(Institute $institute, Course $course)
+    public function edit(BranchOffice $branchOffice, Course $course)
     {
         $this->authorize('tenant-update', $course);
-        $typeCourses = TypeCourse::all();
-        return view('tenant.course.edit', compact('institute', 'course', 'typeCourses'));
+        abort_unless($course->isRegisteredIn($branchOffice), Response::HTTP_NOT_FOUND);
+        $typeCourses = $branchOffice->typecourses;
+        return view('tenant.course.edit', compact('branchOffice', 'course', 'typeCourses'));
     }
 
     /**
      * @param \App\Http\Requests\Tenant\UpdateCourseRequest $request
-     * @param \App\Institute $institute
+     * @param \App\BranchOffice $branchOffice
      * @param \App\Course $course
      * @return \Illuminate\Http\RedirectResponse
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function update(UpdateCourseRequest $request,Institute $institute, Course $course)
+    public function update(UpdateCourseRequest $request, BranchOffice $branchOffice, Course $course)
     {
         $this->authorize('tenant-update', $course);
-        return redirect()->route('tenant.courses.index', $institute)->with(['flash_success' => $request->updateCourse($course)]);
+        abort_unless($course->isRegisteredIn($branchOffice), Response::HTTP_NOT_FOUND);
+        return redirect()->route('tenant.courses.index', $branchOffice)->with(['flash_success' => $request->updateCourse($course)]);
     }
 
     /**
-     * @param \App\Institute $institute
+     * @param \App\BranchOffice $branchOffice
      * @param \App\Course $course
      * @return \Illuminate\Http\RedirectResponse
      * @throws \Illuminate\Auth\Access\AuthorizationException
      * @throws \Exception
      */
-    public function destroy(Institute $institute, Course $course)
+    public function destroy(BranchOffice $branchOffice, Course $course)
     {
         $this->authorize('tenant-delete', $course);
+        abort_unless($course->isRegisteredIn($branchOffice), Response::HTTP_NOT_FOUND);
         $course->delete();
-        return redirect()->route('tenant.courses.index', $institute)->with(['flash_success' => "Curso {$course->name} eliminado con éxito."]);
+        return redirect()->route('tenant.courses.index', $branchOffice)->with(['flash_success' => "Curso {$course->name} eliminado con éxito."]);
     }
 }
